- "mouse gets stuck" bug turned out to be server bug re-introduced with commit  5e78eb1 - Fixed threading bug caused by TextString and global pools - small ammo overlay improvements - fixed strange mouse position behavior when mouse leaves the window (now the last valid mouse position is reported until the mouse enters the window again) - linux build script improvements - improved native interop/ilasm MSBuild targets - windows build script now also copies native server - IPAddressParser now also accepts string 'localhost' - fixed exhaustive lists of enum literal examples having '...' at the end at various places in the console help - improved diagnostic output in case of fatal exceptions - fixed possible use of global pools after the pools have been freed - fixed app crash when socket initialization fails on client connect or server start - fixed incorrect C++ stack-object initialization - "Server has joined the game" message is no longer appearing in the console when a game session is started - removed old C# GL3 code - Linux build script now runs unit tests - Linux: Makefile and Build script now build everything in such a way that the game can be run from the Binaries directory and from within MonoDevelop; Debug builds are not supported yet, as the target path of the Makefile cannot be configured yet - added support for enabling/disabling vsync - added vsync cvar (on by default) - fixed time step frame updates at 60Hz - many UI methods now assert that they're being called on the main thread

+++ Source/Pegasus/Tests/Parsing/BasicParsers/Int16Parser.Tests.cs
@@ -46,13 +46,13 @@
 		[Test]
 		public void Overflow()
 		{
-			Message("1111111111111", String.Format(NumberParser<int>.OverflowMessage, TypeRegistry.GetDescription<short>()));
+			Message("1111111111111", String.Format(NumberParser<short>.OverflowMessage, TypeRegistry.GetDescription<short>()));
 		}
 
 		[Test]
 		public void Underflow()
 		{
-			Message("-1111111111111", String.Format(NumberParser<int>.OverflowMessage, TypeRegistry.GetDescription<short>()));
+			Message("-1111111111111", String.Format(NumberParser<short>.OverflowMessage, TypeRegistry.GetDescription<short>()));
 		}
 
 		[Test]
